[#306] Extra fields on project_update resource
Add "full_name" field with the full name of the user posting the update.

Add "organisation_name" with the name of the organisation the user is
connected to.

Add "organisation" which is a fully working ToOneField to the
organisation of the user.

Add schema objects for "full_name" and "organisation_name" since they
are totally artificial fields.